requirements.txt
PyQt6>=6.6
pyusb>=1.2
